rename to bindable.ts
