+++ packages/greenbox-script/src/bindable.ts
